Reorganize creator files
